refactor: split the main module
create config module and utils directory
