+++ server/config.py
@@ -1,0 +1,5 @@
+import pytz
+
+CPU_SAMPLE_TIME = 4.9
+JOB_CPU_INTERVAL_TIME = 5
+TIME_ZONE = pytz.timezone('Europe/Moscow')
